Collect methods related to Table.
